+++ src/jquery.fancytree.clones.js
@@ -31,13 +31,8 @@
 	/*******************************************************************************
 	 * Private functions and variables
 	 */
-	function _assert(cond, msg) {
-		// TODO: see qunit.js extractStacktrace()
-		if (!cond) {
-			msg = msg ? ": " + msg : "";
-			$.error("Assertion failed" + msg);
-		}
-	}
+
+	var _assert = $.ui.fancytree.assert;
 
 	/* Return first occurrence of member from array. */
 	function _removeArrayMember(arr, elem) {