+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 2.37.1-0 / Unreleased
+  * [Added] #1041 Make assertions more debuggable
 
 # 2.37.0 / 2020-09-11
   * [Changed] #871, #1032 ext-glyph: