+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 2.37.1-0 / Unreleased
+# 2.38.0-0 / Unreleased
   * [Added] #1041 Make assertions more debuggable
   * [Added] #1054 ext-filter tree.updateFilter()